Removed the parsing of angle brackets in the trace macros, it was never going to work properly
It is possible to use top-level template angle brackets in TRACE expressions, but it will garble the output.
examples/trace_example.cpp shows a demonstration:

  TRACE(foo + 1, bar + "_suffix",
        "Template diamond brackets will garble the output. The solution is to put the template expression in brackets "
        "().",
        std::vector<int, std::allocator<int>>{1, 2, 3}, foo + 2);

The parameters std::vector<int, std::allocator<int>>{1, 2, 3} is *two* preprocessor parameters, so the output is:

TRACE at /home/ian/git/uni20/examples/trace_example.cpp:23 : foo + 1 = 43, example_suffix, Template diamond brackets will garble the output. The solution is to put the template expression in brackets ()., std::vector<int = [ 1, 2, 3 ], std::allocator<int>>{1, 2, 3} = 44

where we dropped the final parameter.  It would be possible to give a warning in this case, or even some heuristics to attempt to combine preprocessor parameters, but that seems a bit overkill.  The solution is to disambiguate the expression with brackets, eg

(std::vector<int, std::allocator<int>>{1, 2, 3})

This is anyway necessary for the conditional macros.