Work in progress on refactoring the async core.
We now await on an AsyncTask rather than a raw coroutine_handle, with transfer of ownership being handled by the coroutine promise_type.
This dramatically simplifies some of the buffer code.  However there is a bug somewhere, the examples/async_example.cpp goes into
an infinite loop, depending on the order of the co_await statements.